% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.r
\name{absorptionCompartment}
\alias{absorptionCompartment}
\title{absorptionCompartment}
\usage{
absorptionCompartment(aName = "", isConcentration = FALSE,
  effectsInfo = NULL, isSequential = FALSE, numDosePoints = 1,
  blockName = "")
}
\arguments{
\item{aName}{Absorption variable name}

\item{isConcentration}{Are there volume and concentration parameters}

\item{effectsInfo}{}

\item{isSequential}{Is compartment part of a sequential model}

\item{numDosePoints}{Number of dose points}

\item{blockName}{Name of the structural block}
}
\description{
Creates an absorption compartment
}
\examples{
      abs = absorptionCompartment(aName="Abs", isConcentration=FALSE, 
                             effectsInfo=NULL,
                             numDosePoints="1",

}
