% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{acceptAllEffects}
\alias{acceptAllEffects}
\title{acceptAllEffects}
\usage{
acceptAllEffects(model)
}
\arguments{
\item{model}{PK/PD model}
}
\description{
Accepts all estimates for fixed effects, Sigma and random effects
}
\examples{
     model = acceptAllEffects(model)

}
