% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{methods}
\name{addCategoricalObservation,NlmePmlModel-method}
\alias{addCategoricalObservation,NlmePmlModel-method}
\title{addCategoricalObservation}
\usage{
\S4method{addCategoricalObservation}{NlmePmlModel}(model, name = "C",
  observationName = "CatObs", linkFunction = InvLogit,
  isInhibitory = FALSE, slope = "0", offsetArray = c(),
  dobefore = "", doafter = "", structuralParameters = c(),
  hasRandomEffect = c(), isFrozen = c(), customCodeLines = c())
}
\arguments{
\item{observationName}{Observation name}

\item{linkFunction}{InvLogit|InvProbit|InvLogLog|InvClogClog|InvCustom}

\item{isInhibitory}{Is categorical model inhibitory}

\item{slope}{Slope expression}

\item{offsetArray}{Array of offset  expressions}

\item{dobefore}{Code to execute before the observation}

\item{doafter}{Code to execute after the observation}

\item{structuralParameters}{Array of structural parameters to define}

\item{hasRandomEffect}{Optional array to enable random effects. default = TRUE}

\item{isFrozen}{Optional array to freeze fixed effects. default = FALSE}
}
\description{
Adds a categorical observation to the model
}
\examples{
      model2 = addCategoricalObservation(model,
               observationName="CategoricalObs",
               offsetArray=c("E0 + Emax*C/(C+EC50) + Cat1Constant",
                             "E0 + Emax*C/(C+EC50)"),
               structuralParameters=c("E0","Emax","EC50","Cat1Constant"))

}
