% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addCategoricalObservation}
\alias{addCategoricalObservation}
\title{addCategoricalObservation}
\usage{
addCategoricalObservation(model, ...)
}
\arguments{
\item{observationName}{Observation name}

\item{linkFunction}{Logit|Probit|LogLog|Cloglog|InvCustom}

\item{isInhibitory}{Is the array of offset expressions in
ascending order? (default = FALSE)}

\item{slope}{Slope expression}

\item{offsetArray}{Array of offset  expressions}

\item{dobefore}{Code to execute before the observation}

\item{doafter}{Code to execute after the observation}

\item{structuralParameters}{Array of structural parameters to define}

\item{isFrozen}{Optional array to freeze fixed effects. default = FALSE}

\item{hasRandomEffect}{Optional array to enable random effects. default = TRUE}
}
\description{
Adds a categorical observation to the model
}
\examples{
      model2 = addCategoricalObservation(model,
               observationName="CategoricalObs",
               offsetArray=c("E0 + Emax*C/(C+EC50) + Cat1Constant",
                             "E0 + Emax*C/(C+EC50)"),
               structuralParameters=c("E0","Emax","EC50","Cat1Constant"))

}
