% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{methods}
\name{addContinuousObservation,NlmePmlModel-method}
\alias{addContinuousObservation,NlmePmlModel-method}
\title{addContinuousObservation}
\usage{
\S4method{addContinuousObservation}{NlmePmlModel}(model,
  observationName = "CatObs", effect, structuralParameters = c(),
  hasRandomEffect = c(), isFrozen = c(), customCodeLines = c())
}
\description{
addContinuousObservation
}
