% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addContinuousObservation}
\alias{addContinuousObservation}
\title{addContinuousObservation}
\usage{
addContinuousObservation(model, ...)
}
\arguments{
\item{observationName}{Observation name}

\item{effect}{Specification of residual error model}

\item{structuralParameters}{Optional array of structural parameters to define}

\item{isFrozen}{Optional array to freeze fixed effects. default = FALSE}

\item{hasRandomEffect}{Optional array to enable random effects. default = TRUE}
}
\description{
Adds a continuous observation to the model
}
\examples{
      effect = NlmeResidualEffect(ERR_ADD_MULT,"E",isBQL=TRUE)

      model2 = addContinuousObservation(model,
               observationName="ContinuousObs", effect = effect,
               hasRandomEffect=TRUE)

}
