% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addCountObservation}
\alias{addCountObservation}
\title{addCountObservation}
\usage{
addCountObservation(model, ...)
}
\arguments{
\item{observationName}{Observation name}

\item{slopeExpression}{Hazard slope expression}

\item{expression}{Base hazard expression}

\item{dobefore}{Code to execute before the observation}

\item{doafter}{Code to execute after the observation}

\item{structuralParameters}{Array of structural parameters to define}

\item{isFrozen}{Optional array to freeze fixed effects. default = FALSE}

\item{hasRandomEffect}{Optional array to enable random effects. default = TRUE}
}
\description{
Adds a count observation to the model
}
\examples{
      model2 = addCountObservation(model,
               observationName="CountObs",
               expression="Emax * C / (EC50 + C)",
               structuralParameters=c("Emax","EC50"))

}
