% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addCovariates}
\alias{addCovariates}
\title{addCovariates}
\usage{
addCovariates(model, covariates, effects)
}
\arguments{
\item{model}{PK/PD model}

\item{covariates}{List of covariates variables defined already}

\item{effects}{Covariate/parameter list}
}
\description{
Adds covariates to existing model
}
\examples{
      model=addCovariates(model,
                          c(sex,weight,age),
                          c("EC0"="weight,age",
                          "IC50"= "age,sex"))

}
