% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addEventObservation}
\alias{addEventObservation}
\title{addEventObservation}
\usage{
addEventObservation(model, ...)
}
\arguments{
\item{observationName}{Observation name}

\item{slopeExpression}{Hazard slope expression}

\item{expression}{Base hazard expression}

\item{dobefore}{Code to execute before the observation}

\item{doafter}{Code to execute after the observation}

\item{structuralParameters}{Array of structural parameters to define}

\item{isFrozen}{Optional array to freeze fixed effects. default = FALSE}

\item{hasRandomEffect}{Optional array to enable random effects. default = TRUE}
}
\description{
Adds an event observation to the model
}
\examples{
      model2 = addEventObservation(model,
               observationName="EventObs",
               expression="1 - C/(C+ IC50)",
               structuralParameters=c("IC50"))

}
