% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addExpression}
\alias{addExpression}
\title{addExpression}
\usage{
addExpression(model, ...)
}
\arguments{
\item{model}{PK/PD model}

\item{blockName}{Expression variable name}

\item{codeLine}{Line of PML  defining the expression}

\item{structuralParams}{Optional array of structural parameters to define}

\item{isFrozen}{Optional array to freeze fixed effects. default = FALSE}

\item{hasRandomEffect}{Optional array to enable random effects. default = TRUE for population model}

\item{override}{Override existing definition.(FALSE)}
}
\description{
Adds an expression to the model and optionally override the
definition of an structural parameter.
}
\examples{
     model = addExpression(model,
                      blockName="Cl",
                      structuralParams = c("baseCl","Imax","T50","Gam"),
                      codeLine=" baseCl * (1 - Imax * t^Gam /(t^Gam + T50^Gam))",
                      isFrozen=c(FALSE,TRUE,TRUE,TRUE),
                      override= TRUE)

}
