% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addExtraDose}
\alias{addExtraDose}
\title{addExtraDose}
\usage{
addExtraDose(model, name, doseType, doses)
}
\arguments{
\item{model}{PK/PD model}

\item{name}{Name of the compartment to which the dose is administered}

\item{doseType}{SteadyStateDose|AddlDose}

\item{doses}{List of treatment doses}
}
\description{
Adds an extra dose definition to the model
}
\examples{
     dose = ExtraDoseItem(type=BolusDose,
                          amountColumn="Aa",
                          deltaTimeColumn="ii")

     dose1 = ExtraDoseItem(type=InfusionDose,
                          amount=12,
                          rate=4,
                          deltaTime=1,
                          isSecondDose=TRUE)

     model = addExtraDose(model,name="Aa",
                          doseType=SteadyStateDose,
                          doses=c(dose,dose1) )

}
