% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addLLObservation}
\alias{addLLObservation}
\title{addLLObservation}
\usage{
addLLObservation(model, ...)
}
\arguments{
\item{observationName}{Observation name}

\item{expression}{Logarithm of conditional likelihood}

\item{dobefore}{Code to execute before the observation}

\item{doafter}{Code to execute after the observation}

\item{structuralParameters}{Array of structural parameters to define}

\item{isFrozen}{Optional array to freeze fixed effects. default = FALSE}

\item{hasRandomEffect}{Optional array to enable random effects. default = TRUE}

\item{simulationCode}{Array of simulation code lines}
}
\description{
Adds an LL observation to the model
}
\examples{
      model2 = addLLObservation(model,
               observationName="iObs",
               expression="-E +iObs*log(E) - lgamm(iObs+1)",
               customCodeLines=c("double(prob,i,u)"),
               simulationCode=c("u = unif()",
                                "prob = 0",
                                "i = 0",
                                "prob = prob + exp(-E + i*log(E) -lgamm(i+1))",
                                "while(u >= prob){",
                                "    prob = prob + exp(-E + i * log(E) -lgamm(i+1))",
                                "    i = i + 1",
                                " }",
                                " iObs = i"))

}
