% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addParameter}
\alias{addParameter}
\title{addParameter}
\usage{
addParameter(model, ...)
}
\arguments{
\item{model}{PK/PD model}

\item{name}{Name of the structural parameter}

\item{fixedEffName}{Name to use for fixed effects}

\item{randomEffName}{Name to use for random effects}

\item{hasRandomEffect}{Does the parameter have a random effect? (default:TRUE)}

\item{style}{Parameter style:
LogNormal, Normal, Combination, Log, Logit, Custom}

\item{initialValue}{Initial value for the fixed effect}

\item{lowerBound}{Lower limit for the fixed effect}

\item{upperBound}{Upper limit for the fixed effect}

\item{isFrozen}{Freeze fixed effect? (default : FALSE)}

\item{ranEffInitValue}{Initial value for the random effect}

\item{code}{For Custom style, PML code to override the definition}

\item{extraCode}{Extra lines of code that relates to this parameter}
}
\description{
Adds a structural parameter to the model and optionally
override the definition of an structural parameter.
}
\examples{
     model = addParameter(model,
                          name="baseCl",
                          style=LogNormal)

}
