% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addProcedure}
\alias{addProcedure}
\title{addProcedure}
\usage{
addProcedure(model, ...)
}
\arguments{
\item{model}{PK/PD model}

\item{blockName}{Procedure block name}

\item{codeLines}{Line of PML code to insert into model}

\item{structuralParams}{Optional array of structural parameters to define}

\item{isFrozen}{Optional array to freeze fixed effects. default = FALSE}

\item{hasRandomEffect}{Optional array to enable random effects. default = TRUE for population model}

\item{override}{Override existing definition.(FALSE)}
}
\description{
Adds a procedure block to the model and optionally override the definition of
structural parameters.
}
\examples{
     model = addProcedure(model,
                          blockName="Del",
                          structuralParams = c("Del"),
                          codeLines=c("Ka=Ke+Del"),
                          isFrozen=c(FALSE),
                          override= TRUE)

}
