% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addReset}
\alias{addReset}
\title{addReset}
\usage{
addReset(model, low, hi)
}
\arguments{
\item{model}{PK/PD model}

\item{low}{Lower value of reset range}

\item{hi}{Upper value of reset range}
}
\description{
Adds reset dose instructions to the model
}
\examples{

}
