% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{addSecondary}
\alias{addSecondary}
\title{addSecondary}
\usage{
addSecondary(.Object, name, definition, unit = "")
}
\arguments{
\item{name}{Name of the secondary parameter}

\item{definition}{Definition of secondary variable}

\item{unit}{Optional units}

\item{model}{PK/PD model}
}
\description{
Adds a secondary parameter to model definition
}
\examples{

    model = addSecondary(model,"Spc_Param","log(2)/tvKe")
    model = addSecondary(model,"Tmax",
                         "CalcTMax(tvA,tvAlpha,tvB,tvBeta,C,Gamma)")

}
