% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{addTablesToColumnMapping}
\alias{addTablesToColumnMapping}
\title{addTablesToColumnMapping}
\usage{
addTablesToColumnMapping(model, simParams, filename)
}
\arguments{
\item{model}{A PK/PD model}

\item{simParams}{Simulation parameters}

\item{filename}{Name of simulation table file}
}
\description{
Add simulation tables to the mapping
}
\examples{
       addTablesToColumnMapping(model,vpcParam,filename)

}
