% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\name{addToErrorModel}
\alias{addToErrorModel}
\title{addToErrorModel}
\usage{
addToErrorModel(model, effectsList)
}
\arguments{
\item{effectsList}{}
}
\description{
Add to the NLME Error model
}
\examples{

}
