% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{blankmodel}
\alias{blankmodel}
\title{Method to create an empty model}
\usage{
blankmodel(modelName = "", workingDir = "", isPopulation = TRUE)
}
\arguments{
\item{modelName}{Visual name for the model}

\item{workingDir}{working directory to run the model}
}
\description{
Method to create an empty model
}
\examples{

      model = model()

}
