% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_param.r
\name{categoricalCovariate}
\alias{categoricalCovariate}
\title{categoricalCovariate}
\usage{
categoricalCovariate(name = "", categories = c(),
  categoryNames = c(), direction = Backward)
}
\arguments{
\item{name}{Name of the covariate}

\item{categories}{Numerical values for each category}

\item{direction}{Curve fitting method: Forward|Backward|Interpolate}

\item{categorNames}{Names of each category}
}
\description{
Convenience method to create a categorical covariate
}
\examples{
      sex=categoricalCovariate("sex",
                       c(1,2),
                       c("female","male"),
                       direction=Forward)

}
