% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.r
\name{centralCompartment}
\alias{centralCompartment}
\title{centralCompartment}
\usage{
centralCompartment(aName = "", vName = "", cName = "",
  isConcentration = TRUE, isSequential = FALSE, numDosePoints = 0,
  blockName = "")
}
\arguments{
\item{aName}{Absorption variable name}

\item{vName}{Volume variable name}

\item{cName}{Observation variable name}

\item{isConcentration}{Are there volume and concentration parameters}

\item{isSequential}{Is compartment part of a sequential model}

\item{numDosePoints}{Number of dose points}

\item{blockName}{Name of the structural block}
}
\description{
Creates a central compartment
}
\examples{
    cenComp = centralCompartment(aName="Abs1",
                       vName="Vol",
                       cName="Conc",
                       isConcentration=TRUE,
                       isSequential=FALSE,
                       numDosePoints="2",
                       blockName="Cent1")

}
