% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.r
\docType{class}
\name{CompartmentDoseType-class}
\alias{CompartmentDoseType-class}
\alias{CompartmentDoseType}
\title{CompartmentDoseType}
\arguments{
\item{dosetype}{Dose type: 1=NoDose, 2=OneDose, 3=TwoDoses, 4=SplitDoses}
}
\description{
Class represents type of dose
}
\examples{
       CompartmentDoseType(doseType="1")

}
