% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.r
\docType{class}
\name{CompartmentType-class}
\alias{CompartmentType-class}
\alias{CompartmentType}
\title{CompartmentType}
\arguments{
\item{style}{Compartment style: 1=Central, 2=Absorption, 3=Peripheral, 4=Elimination}
}
\description{
Class represents type of compartment
}
\examples{
      CompartmentType(styl="1")

}
