% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_param.r
\name{continuousCovariate}
\alias{continuousCovariate}
\title{continuousCovariate}
\usage{
continuousCovariate(name = "", centerValue = 0, isPositive = FALSE,
  direction = Backward, continuousType = CovarNumber)
}
\arguments{
\item{name}{Name of the covariate}

\item{centerValue}{Covariate centering value}

\item{isPositive}{Are the covariate values all positive}

\item{direction}{Curve fitting method: Forward|Interpolate|Backward}

\item{continuousType}{Type of value to use for the centering value:
CovarNumber|CovarMean|CovarMedian}
}
\description{
continuousCovariate
}
