% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{copyModel}
\alias{copyModel}
\title{Copies a model into a new object, new working directory is created
and all files are copied into it}
\usage{
copyModel(model, modelName = "", workingDir = "")
}
\description{
Copies a model into a new object, new working directory is created
and all files are copied into it
}
