% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{copyPhoenixFiles}
\alias{copyPhoenixFiles}
\title{copyPhoenixFiles}
\usage{
copyPhoenixFiles(model, dataset)
}
\arguments{
\item{model}{PK/PD model}

\item{dataset}{Dataset to run simulation/fitting on}
}
\description{
Copy Phoenix model/data files into the model directory
}
