% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{covariateEffect}
\alias{covariateEffect}
\title{covariateEffect}
\usage{
covariateEffect(.Object, covariateName, parameterName)
}
\arguments{
\item{covariateName}{Name of the covariate}

\item{model}{A PK/PD model}

\item{variableName}{Name of the model variable}
}
\value{
effect style    One of: EnableEffect|DisableEffect|PlusOneEffect
}
\description{
Lists style for a covariate/variable
}
\examples{
    covariateEffect(pkpdmodel,"age","Cl")

    covariateEffect(pkpdmodel,"age","Cl") = PlusOneEffect

}
