% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{createInitialDoseMapping}
\alias{createInitialDoseMapping}
\title{createInitialDoseMapping}
\usage{
createInitialDoseMapping(model, doseInput)
}
\arguments{
\item{model}{A PK/PD model}

\item{doseInput}{Column names in dose file}
}
\description{
Creates a mapping between model variable and dose columns
}
\examples{
   createInitialDoseMapping(model,doseInput)

}
