% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{createInitialMapping}
\alias{createInitialMapping}
\title{createInitialMapping}
\usage{
createInitialMapping(model, inputData)
}
\arguments{
\item{model}{A PK/PD model}

\item{InputData}{Column name}
}
\description{
Creates a mapping between model variable and dataset columns
}
\examples{
   createInitialMapping(model,InputData)

}
