% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{createInitialParamsMapping}
\alias{createInitialParamsMapping}
\title{createInitialParamsMapping}
\usage{
createInitialParamsMapping(model, paramsInput)
}
\arguments{
\item{model}{A PK/PD model}

\item{paramsInput}{Parameter column names}
}
\description{
Creates a mapping between model variable and parameter columns
}
\examples{
   createInitialParamsMapping(model,doseInput)

}
