% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{createInitialRandEffectMapping}
\alias{createInitialRandEffectMapping}
\title{createInitialRandEffectMapping}
\usage{
createInitialRandEffectMapping(model, randEffectInput)
}
\arguments{
\item{model}{A PK/PD model}

\item{randEffectInput}{Random effect data}
}
\description{
Creates a mapping between model variables and random effects columns
}
\examples{
    createInitialRandEffectMapping(model,randEffectInput)

}
