% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{defaultDataset}
\alias{defaultDataset}
\title{defaultDataset}
\usage{
defaultDataset(model, inputFile)
}
\arguments{
\item{model}{A PK/PD model}

\item{input}{Data frame with subject/observation data}
}
\value{
dataset  NLME dataset class that links model to input
}
\description{
Formats column names of a data table and writes it out
to given file name
}
\examples{
  dataset = defaultDataset(model,input)

}
