% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{deleteSecondary}
\alias{deleteSecondary}
\title{deleteSecondary}
\usage{
deleteSecondary(.Object, name)
}
\arguments{
\item{name}{Name of the secondary parameter}

\item{model}{PK/PD model}
}
\description{
Deletes a secondary parameter from the model
}
