% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_param.r
\docType{data}
\name{Diagonal}
\alias{Diagonal}
\title{Diagonal random effects}
\format{An object of class \code{numeric} of length 1.}
\usage{
Diagonal
}
\description{
Diagonal random effects
}
\keyword{datasets}
