% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{editModel}
\alias{editModel}
\title{editModel}
\usage{
editModel(model, dataset)
}
\arguments{
\item{model}{A PK/PD model}

\item{dataset}{Dataset}
}
\description{
Allows user to edit PML text and column definition.  It will return a new textual model
containing the edited model.
}
\examples{

      newModel = editModel(model,dataset)

}
