% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_compartment.r
\docType{class}
\name{EffectsCompartment-class}
\alias{EffectsCompartment-class}
\alias{EffectsCompartment}
\title{EffectsCompartment}
\arguments{
\item{flowInName}{}

\item{ceName}{Input variable name for port1}

\item{ke0Name}{Input variable name for port2}
}
\description{
Class represents an NLME Effects compartment
}
\examples{

}
