% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.r
\name{eliminationCompartment}
\alias{eliminationCompartment}
\title{eliminationCompartment}
\usage{
eliminationCompartment(aName = "", vName = "", cName = "",
  feName = "", isSequential, isFractionExcreted, numDosePoints = 0,
  blockName = "")
}
\arguments{
\item{aName}{Absorption variable name}

\item{vName}{Volume variable name}

\item{cName}{Observation variable name}

\item{feName}{Fraction Excreted variable name}

\item{isSequential}{Is compartment part of a sequential model}

\item{isFractionExcreted}{Does compartment include a fraction excreted parameter}

\item{numDosePoints}{Number of dose points}

\item{blockName}{Name of the structural block}
}
\description{
Creates an elimination compartment
}
\examples{
    elimComp = eliminationCompartment(aName="Abs0",
                           vName="Vol0",
                           cName="Conc0",
                           feName="FrEx",
                           isSequential=FALSE,
                           isFractionExcreted=TRUE,
                           numDosePoints="2",
                           blockName="Elim1")

}
