% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_compartment.r
\docType{class}
\name{EmaxCompartment-class}
\alias{EmaxCompartment-class}
\alias{EmaxCompartment}
\title{EmaxCompartment}
\arguments{
\item{cName}{Concentration variable name}

\item{c50Name}{IC50 name}

\item{firstName}{First emax parameter name}

\item{gammaName}{Gamma variable name}

\item{maxName}{Imax variable name}

\item{checkBaseline}{Does model have a baseline response}

\item{checkFractional}{Is the model fractional}

\item{checkInhibitory}{Is the model inhibitory}

\item{checkSigmoid}{Is the model sigmoidal}
}
\description{
Class represents an NLME Emax compartment
}
\examples{
     emaxComp = EmaxCompartment(checkBaseline=TRUE,
                                checkFractional = TRUE,
                                checkInhibitory = TRUE,
                                checkSigmoid = TRUE,
                                blockName = "E")

}
