% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.r
\name{estimatesUI}
\alias{estimatesUI}
\title{estimatesUI}
\usage{
estimatesUI(model, subjectIds, host = NULL, dataset = NULL)
}
\arguments{
\item{model}{PK/PD model}

\item{subjectIds}{Subject IDs}

\item{host}{Optional host parameter if model needs to be compiled}

\item{dataset}{Optional dataset parameter}
}
\description{
User Interface to examine the model and evaluate estimates for fixed effects
}
\examples{
   host = NlmeParallelHost(sharedDirectory=Sys.getenv("NLME_ROOT_DIRECTORY"),
                           parallelMethod=NlmeParallelMethod("LOCAL_MPI"),
                           hostName="MPI",
                           numCores=4)
   input=read.csv("D:/SDCard/NlmeInstall_04_30_18/Pml/pk01cov3.csv")
   model = pkmodel(numComp=1,
                   isPopulation=TRUE,
                   absorption = Intravenous,
                   parameterization = Clearance,
                   modelName="PK01Model",
                   isTlag = FALSE,
                   hasEliminationComp = FALSE,
                   isClosedForm = TRUE)

   dataset=NlmeDataset(model@modelInfo@workingDir)
   initColMapping(model)=input
   modelColumnMapping(model)=c(ID="xid", CObs="conc",A1="dose")
   writeDefaultFiles(model,dataset)
   estimatesUI(model,unique(input$id),host)

   estimates = getInitialEstimates(model)
   print(estimates)
   initFixedEffects(model) = estimates

}
