% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing.r
\docType{class}
\name{ExtraDoseDataType-class}
\alias{ExtraDoseDataType-class}
\alias{ExtraDoseDataType}
\title{ExtraDoseDataType}
\arguments{
\item{type}{ValueType|ColumnType}

\item{value}{Value of dosing parameter}

\item{column}{Name of column containing dose information}
}
\description{
ExtraDoseDataType
}
\examples{

}
