% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing.r
\docType{class}
\name{ExtraDoseItem-class}
\alias{ExtraDoseItem-class}
\alias{ExtraDoseItem}
\title{ExtraDoseItem}
\arguments{
\item{name}{Dose point name}

\item{type}{BolusDose|InfusionDose}

\item{amount}{Optional dose amount}

\item{rate}{Optional dose rate}

\item{deltaTime}{Optional delta time}

\item{amountColumn}{Optional column name containing dose amount data}

\item{rateColumn}{Optional column name containing dose rate data}

\item{deltaTimeColumn}{Optional column name containing dose delta time}

\item{isSecondDose}{Use second dose point on compartment}
}
\description{
ExtraDoseItem
}
\examples{

dose = ExtraDoseItem(type=BolusDose,
                     amountColumn="Aa",
                     deltaTimeColumn="ii")

dose1 = ExtraDoseItem(type=InfusionDose,
                     amount=12,
                     rate=4,
                     deltaTime=1,
                     isSecondDose=TRUE)

}
