% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing.r
\docType{class}
\name{ExtraDoseOption-class}
\alias{ExtraDoseOption-class}
\alias{ExtraDoseOption}
\title{ExtraDoseOption}
\arguments{
\item{name}{Dose name}

\item{doseType}{SteadyStateDose|AddlDose}

\item{doses}{List of treatment doses}
}
\description{
Additional dosing information
}
\examples{

dose = ExtraDoseItem(type=BolusDose,
                     amountColumn="Aa",
                     deltaTimeColumn="ii")

dose1 = ExtraDoseItem(type=InfusionDose,
                     amount=12,
                     rate=4,
                     deltaTime=1,
                     isSecondDose=TRUE)

dosePoint = ExtraDoseOption(name="SteadyState",
                          doseType=SteadyStateDose,
                          doses=c(dose,dose1) )

model = addExtraDose(model,dosePoint)


}
