% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{generateColumnMappingFile}
\alias{generateColumnMappingFile}
\title{Method to create column mapping file}
\usage{
generateColumnMappingFile(model, fileName)
}
\arguments{
\item{model}{Model to be mapped}

\item{fileName}{Name of file containing column mapping}
}
\description{
Method to create column mapping file
}
\examples{

}
