% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{generateObjectModel}
\alias{generateObjectModel}
\title{generateObjectModel}
\usage{
generateObjectModel(obj)
}
\description{
Generates a connection table of model objects that represents
the built-in model
}
