% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_param.r
\name{getCovariate}
\alias{getCovariate}
\title{getCovariate}
\usage{
getCovariate(model, varname)
}
\arguments{
\item{model}{Name of model}

\item{varname}{Name of the covariate variable}
}
\description{
Retrieve a covariate from a list
}
\examples{

}
