% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.r
\name{getInitialEstimates}
\alias{getInitialEstimates}
\title{getInitialEsimates}
\usage{
getInitialEstimates()
}
\description{
Returns values from initial estimates shiny App.  
Returned value can be used to set initial estimates in RsNlme
}
\examples{
     ...
     estimatesUI(model,unique(input$ID),host)

     estimates = getInitialEstimates()

     initFixedEffects(model) = estimates

}
