% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{getThetas}
\alias{getThetas}
\title{getThetas}
\usage{
getThetas(.Object)
}
\arguments{
\item{model}{PK/PD model}
}
\description{
Gets names and values of Thetas by parsing PML model for fixef()
}
\examples{
      getThetas(pkpdmodel)
            tvKa    tvV   tvCl tvEC50   tvE0 tvEmax
             "1"    "1"    "1"  "1"     "1"    "1"

}
