% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_indirect.r
\docType{class}
\name{IndirectCompartment-class}
\alias{IndirectCompartment-class}
\alias{IndirectCompartment}
\title{IndirectCompartment}
\arguments{
\item{cName}{Concentration variable name}

\item{emaxName}{emax variable name}

\item{ec50Name}{ec50 variable name}

\item{kinName}{kin variable name}

\item{koutName}{kout variable name}

\item{gammaName}{gamma variable name}

\item{sName}{}

\item{isExponent}{Is there an exponent Gamma added to the effect statement}

\item{isBuildup}{Is the response formation (TRUE) or degradation (FALSE) 
affected by increased concentrations}

\item{type}{LimitedStimulation|InfiniteStimulation|
LimitedInhibition|InverseInhibition|
LinearStimulation|LogLinearStimulation}
}
\description{
Class represents an NLME Indirect compartment
}
\examples{

}
