% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{methods}
\name{initColMapping<-,NlmePmlModel-method}
\alias{initColMapping<-,NlmePmlModel-method}
\title{initColMapping}
\usage{
\S4method{initColMapping}{NlmePmlModel}(.Object) <- value
}
\arguments{
\item{model}{A PK/PD model with the variables to be mapping}

\item{dataset}{Input dataset}
}
\description{
Tries to map the input dataset columns to model variables
}
