% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{initDoseColMapping<-}
\alias{initDoseColMapping<-}
\title{initDoseColMapping}
\usage{
initDoseColMapping(.Object) <- value
}
\arguments{
\item{model}{PK/PD model}

\item{doseSpreadsheet}{Data frame containing the dose information}
}
\description{
Maps columns in the dosing spreadsheet to model dose variables by string matching
}
\details{
It is used when dose values are supplied in a separate input file
}
\examples{
    doseInput=read.csv("PKPD_Data.csv")
    initDoseColMapping(pkpdmodel)=doseInput

    print(modelDoseMapping(pkpdmodel))
    print(doseNames(pkpdmodel))
    modelDoseMapping(pkpdmodel)=c(Aa="Dose")

    Dose  values and  mappings are supplied in cols2.txt and
    data2.txt files to the engine

}
