% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{initFixedEffects<-}
\alias{initFixedEffects<-}
\title{initFixedEffects}
\usage{
initFixedEffects(.Object) <- value
}
\arguments{
\item{model}{PK/PD model}
}
\description{
Assigns the initial values for fixed effects
It should be used after all model components have been created.
}
\examples{
      initFixedEffects(model) = c(tvIC50=0.987,tvE0=322,tvImax=0)

}
