% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{initFixedEffects}
\alias{initFixedEffects}
\title{initFixedEffects}
\usage{
initFixedEffects(.Object)
}
\arguments{
\item{.Object}{PK/PD model}
}
\description{
Prints/sets the initial values for fixed effects.
}
\examples{
      initFixedEffects(pkpdmodel)
            tvKa    tvV   tvCl tvEC50   tvE0 tvEmax
             "1"    "1"    "1"  "1"     "1"    "1"

      initFixedEffects(model) = c(tvIC50=0.987,tvE0=322,tvImax=0)

}
