% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_eff.r
\name{initializeRandomEffectsBlock}
\alias{initializeRandomEffectsBlock}
\title{initializeRandomEffectsBlock}
\usage{
initializeRandomEffectsBlock(.Object)
}
\arguments{
\item{model}{PK/PD model}
}
\description{
Initializes random effects structure from structural parameters
}
