% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{initParamsMapping<-}
\alias{initParamsMapping<-}
\title{initParamsMapping}
\usage{
initParamsMapping(.Object) <- value
}
\arguments{
\item{model}{PK/PD model}

\item{paramsInput}{Data frame containing fixed effects information}
}
\description{
Maps columns in the fixed effects dataframe  to model  variables
by string matching
}
\details{
It is used when initial values for fixed effects are supplied from
a previous fitting run.
}
\examples{
    paramsInput=read.csv("params.csv")
    head(paramsInput)
      Parameter Initial Lower Upper
      tvKa        3       1     18
      tvV         4      -9     44
      tvCl       NA      NA     NA

    initParamsMapping(model)=paramsInput

    modelParamsMapping(model)
    dataset = updateDatasetParams(model,dataset)

    Fixed effect values and column mappings are supplied in cols3.txt and
    data3.txt files to the engine

}
