% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_eff.r
\name{initRandomEffects}
\alias{initRandomEffects}
\title{initRandomEffects}
\usage{
initRandomEffects(.Object)
}
\arguments{
\item{model}{PK/PD model}

\item{values}{Random effect values specified as a PML string OR one or
more sets of quadruplet values}
}
\description{
Initializes the random effect statement. It should be used after all model components have been created.
}
\examples{
     initRandomEffects(model)=c(PMLStringForRanef)

     initRandomEffects(model)=c(BlockOrDiagonal,
                                isFrozen,
                                listOfVariables,
                                listOfValues)

     initRandomEffects(model)=c(Block,
                                FALSE,
                                "nV,nCl,nKa,nV2",
                                "0.2, 0, 0.2, 0, 0, 0.2, 0, 0, 0, 0.1")

     initRandomEffects(model)=c(Diagonal, FALSE,"nV,nCl","0.1, 0.02")

     initRandomEffects(model)=c(Diagonal, FALSE,"nV,nCl","0.1, 0.02",
                                Block, TRUE, "nCl2,nV2","0.2, 0, 0.2")

     initRandomEffects(model)= c("block(nCl, nV)(freeze) = 
                               c(0.2, 0, 0.2), diag(nKa)(freeze) = 
                               c(0.2), diag(nV2) = 
                               c(0.1), diag(nV3, nCl2, nCl3) = 
                               c(1, 1, 1)")

}
