% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{initRandParamsMapping<-}
\alias{initRandParamsMapping<-}
\title{initRandParamsMapping}
\usage{
initRandParamsMapping(.Object) <- value
}
\arguments{
\item{model}{PK/PD model}

\item{paramsInput}{Data frame containing random effects information}
}
\description{
Maps columns in the random effects dataframe  to model  variables
by string matching
}
\details{
It is used when initial values for random effects are supplied from
a previous fitting run.
}
\examples{
    paramsInput=read.csv("ranparams.csv")
    OR
    source("dmp.txt")
    paramsInput=dmp.txt$coefficients$random[[1]]

    head(paramsInput)

             nV        nCl
          -0.617345 -0.3034520
           0.263520  0.5110880
          -0.439680 -0.3502870

    initRandParamsMapping(model)=paramsInput

    modelRandParamsMapping(model)
    dataset = updateDatasetRandParams(model,dataset)

    Fixed effect values and column mappings are supplied in cols4.txt and
    data4.txt files to the engine

}
