% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_linear.r
\docType{class}
\name{LinearCompartment-class}
\alias{LinearCompartment-class}
\alias{LinearCompartment}
\title{LinearCompartment}
\arguments{
\item{cName}{Concentration variable name}

\item{alphaName}{Alpha variable name}

\item{betaName}{Beta variable name}

\item{gammaName}{Gamma variable name}

\item{type}{LinearAlpha|LinearBeta|LinearGamma}
}
\description{
Class represents an NLME Linear compartment
}
\examples{
     linearComp = LinearCompartment(
                          type = LinearAlpha,
                          blockName = "E")

}
