% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{linearmodel}
\alias{linearmodel}
\title{linearModel}
\usage{
linearmodel(isPopulation = TRUE, type = LinearAlpha,
  covariates = c(), modelName = "", workingDir = "")
}
\arguments{
\item{isPopulation}{Population vs Individual models}

\item{type}{Type of Linear model
LinearAlpha|LinearBeta|LinearGamma}

\item{covariates}{List of covariates defined for the model}

\item{modelName}{Visual name for the model}

\item{workingDir}{Working directory to run the model}
}
\description{
Method to create a Linear PML model
}
\examples{
       model=linearmodel()

       OR

       model=linearmodel(type=LinearBeta,
                         modelName="Linear",
                         workingDire="./"))

       print(attr(model,"statements"))

       "test(){"
       "	covariate(C)"
       "	E = Alpha + Beta*C"
       "	error(EEps = 1)"
       "	observe(EObs(C) = E + EEps)"
       "	stparm(Alpha = tvAlpha * exp(nAlpha))"
       "	stparm(Beta = tvBeta * exp(nBeta))"
       "	fixef(tvAlpha = c(, 1, ))"
       "	fixef(tvBeta = c(, 1, ))"
       "	ranef(diag(nAlpha, nBeta) = c(1, 1))"
       "}"

}
