% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{loadModel}
\alias{loadModel}
\title{loadModel}
\usage{
loadModel(directory)
}
\arguments{
\item{directory}{Directory where the model was saved}
}
\description{
Loads a previously saved model from disk and sets cwd to Model's working directory
}
\examples{
      model=loadModel(directory)

}
