% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{lookupColname}
\alias{lookupColname}
\title{lookupColname}
\usage{
lookupColname(mapping, vname)
}
\arguments{
\item{mapping}{Mapping list}

\item{vname}{Variable name}
}
\description{
Return column names
}
\examples{
   lookupColname(mapping,vname)

}
