% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{methods}
\name{mixedFirstOrderZeroOrderAbsorption,NlmePmlModel-method}
\alias{mixedFirstOrderZeroOrderAbsorption,NlmePmlModel-method}
\title{mixedFirstOrderZeroOrderAbsorption}
\usage{
\S4method{mixedFirstOrderZeroOrderAbsorption}{NlmePmlModel}(model,
  doseInfo, absDoseInfo)
}
\arguments{
\item{model}{PK/PD model to apply absorption model to}

\item{doseInfo}{Dose information for central compartment}

\item{absDoseInfo}{Dose information for absorption compartment}
}
\description{
Creates a mixed 1st, zeroth order absorption
dosing regimen for a given model
}
\examples{
      dosePoint= DosePoint(isZeroOrderAbsorption = DurationDose,
                           durationExpression = "DUR",
                           isTlag = FALSE,
                           isBioavail = TRUE,
                           bioavailExpression = "logitF1;1-ilogit(logitF1)")

      absDosePoint= DosePoint(isZeroOrderAbsorption = NoDose,
                              isBioavail = TRUE,
                              bioavailExpression = "logitF1;ilogit(logitF1)")

      model2=mixedFirstOrderZeroOrderAbsorption(model,dosePoint,absDosePoint)
      print(model2)

}
