% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{methods}
\name{modelColumnMapping,NlmePmlModel-method}
\alias{modelColumnMapping,NlmePmlModel-method}
\title{Lists mapping between model variables and data columns}
\usage{
\S4method{modelColumnMapping}{NlmePmlModel}(.Object)
}
\description{
Lists mapping between model variables and data columns
}
\examples{
      modelColumnMapping(model)=c(id="Subject", CObs="Conc",A1="Amount")j

}
