% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{modelDoseMapping<-}
\alias{modelDoseMapping<-}
\title{Sets mapping between model variables and data columns}
\usage{
modelDoseMapping(.Object) <- value
}
\arguments{
\item{List}{of variable/column pairs}
}
\description{
Sets mapping between model variables and data columns
}
\examples{
      modelColumnMapping(model)=c(id="Subject", CObs="Conc",A1="Amount")j

}
