% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{modelObservationVariables}
\alias{modelObservationVariables}
\title{modelObservationVariables}
\usage{
modelObservationVariables(model)
}
\arguments{
\item{model}{PK/PD model}
}
\description{
Returns observation variables defined for this model
}
\examples{

      obsvars = modelObservationVariables(pkmodel)

}
