% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{modelParamsMapping}
\alias{modelParamsMapping}
\title{modelParamsMapping}
\usage{
modelParamsMapping(.Object)
}
\description{
Lists mapping between model fixed effects and input columns
}
\examples{
      modelParamsMapping(model)

}
