% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{modelRandParamsMapping}
\alias{modelRandParamsMapping}
\title{modelRandParamsMapping}
\usage{
modelRandParamsMapping(.Object)
}
\description{
Lists mapping between model random effects and input columns
}
\examples{
      modelRandParamsMapping(model)

}
