% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{modelVariableNames}
\alias{modelVariableNames}
\title{modelVariableNames}
\usage{
modelVariableNames(model)
}
\arguments{
\item{model}{A PK/PD model}
}
\description{
Return model variable names
}
\examples{
   modelVariableNames(model)

}
