% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\docType{class}
\name{NlmeColumnMap-class}
\alias{NlmeColumnMap-class}
\alias{NlmeColumnMap}
\title{NlmeColumnMap}
\arguments{
\item{variableName}{Model variable name}

\item{columnName}{Corresponding column name in the dataset}
}
\description{
Class represents map between a model variable and a dataset column
}
\examples{
   mapping = NlmeColumnMap(variableName,columnName)

}
