% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\docType{class}
\name{NlmeColumnMapping-class}
\alias{NlmeColumnMapping-class}
\alias{NlmeColumnMapping}
\title{NlmeColumnMapping}
\description{
Class represents mapping list between model variables and dataset columns
}
\examples{
   mapping = NlmeColumnMapping("list")

}
