% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.r
\docType{class}
\name{NlmeCompartment-class}
\alias{NlmeCompartment-class}
\alias{NlmeCompartment}
\title{NlmeCompartment}
\arguments{
\item{aName}{Absorption variable name}

\item{cName}{Observation variable name}

\item{vName}{Volume variable name}

\item{feName}{Fraction Excreted variable name}

\item{inputName}{Compartment name}

\item{hasVolume}{Does compartment have volume variable}

\item{isConcentration}{Is compartment a concentration compartment}

\item{isFractionExcreted}{Does compartment include a fraction excreted parameter}

\item{isSequential}{Is compartment part of a sequential model}

\item{type}{Central|Elimination|Peripheral}
}
\description{
Class represents an NLME compartment
}
\examples{
    centralCompartment=NlmeCompartment(inputVariableName="A1",
                         aName="",
                         vName="V",
                         cName="C",
                         type=Central)

}
