% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_param.r
\docType{class}
\name{NlmeCovarItem-class}
\alias{NlmeCovarItem-class}
\alias{NlmeCovarItem}
\title{NlmeCovarItem}
\arguments{
\item{name}{Name of covariate}

\item{value}{Covariate value}
}
\description{
Class represents an NLME Covariate parameter
}
\examples{
      NlmeCovarItem(name,value)

}
