% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\docType{class}
\name{NlmeDoseMapping-class}
\alias{NlmeDoseMapping-class}
\alias{NlmeDoseMapping}
\title{NlmeDoseMapping}
\description{
Class represents mapping list between model variables and dose columns
}
\examples{
   mapping = NlmeDoseMapping("list")

}
