% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{class}
\name{NlmeEmaxParameters-class}
\alias{NlmeEmaxParameters-class}
\alias{NlmeEmaxParameters}
\title{NlmeEmaxParameters}
\arguments{
\item{checkBaseline}{Model has a baseline response}

\item{checkFractional}{Model is fractional}

\item{checkInhibitory}{Model is inhibitory}

\item{checkSigmoid}{Model is sigmoidal}
}
\description{
Class represents an NLME/PML Emax model parameters
}
\examples{
     NlmeEmaxParameters(checkBaseline=TRUE)

}
