% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.r
\docType{class}
\name{NlmeExpression-class}
\alias{NlmeExpression-class}
\alias{NlmeExpression}
\title{NlmeExpression}
\arguments{
\item{blockName}{Name of the expression block and LHS of assignment}

\item{structuralParams}{List of structural parameters to define}

\item{codeLine}{PML statement for expression}
}
\description{
Defines a PML Expression block with optional structural parameters
}
\examples{

expr1 = NlmeExpression(blockName = "C2",
                       structuralParams=c("Vp"),
                       codeLine=c("Ap / Vp"))

addCompartment(model1) = expr1

}
