% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{class}
\name{NlmeModelAbsorption-class}
\alias{NlmeModelAbsorption-class}
\alias{NlmeModelAbsorption}
\title{NlmeModelAbsorption}
\arguments{
\item{absorpType}{Intravenous|Extravascular}
}
\description{
Class represents an NLME/PML model absorption
}
\examples{
      NlmeModelAbsorption(PARAM_EXTRAVASCULAR)

}
