% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{class}
\name{NlmeModelParameterization-class}
\alias{NlmeModelParameterization-class}
\alias{NlmeModelParameterization}
\title{NlmeModelParameterization}
\arguments{
\item{paramType}{One of Micro|Clearance|Macro|Macro1}
}
\description{
Class represents an NLME/PML model parameterization
}
\examples{
      NlmeModelParameterization(PARAM_CLEARANCE)

}
