% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{class}
\name{NlmeModelType-class}
\alias{NlmeModelType-class}
\alias{NlmeModelType}
\title{NlmeModelType}
\arguments{
\item{modelType}{}
}
\description{
Class represents an NLME/PML model type
}
\examples{
      NlmeModelType(PK)

}
