% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\docType{class}
\name{NlmeParamsMapping-class}
\alias{NlmeParamsMapping-class}
\alias{NlmeParamsMapping}
\title{NlmeParamsMapping}
\description{
Class represents mapping list between model variables and data columns
}
\examples{
   mapping = NlmeParamsMapping("list")

}
