% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{class}
\name{NlmePkParameters-class}
\alias{NlmePkParameters-class}
\alias{NlmePkParameters}
\title{NlmePkParameters}
\arguments{
\item{parameterization}{Taken from NlmeModelParameterization}

\item{absorption}{Taken from NlmeModelAbsorption}

\item{numCompartments}{Number of compartments}

\item{isTlag}{Does dose have a time lag}

\item{hasEliminationComp}{Is there data available for an elimination compartment}

\item{isFractionExcreted}{Does compartment include a fraction excreted parameter}

\item{isMichaelisMenten}{Is this a Michaelis-Menten model}

\item{isSaturating}{Is the elimination rate equal to the absorption rate}

\item{infusionAllowed}{Is infusion allowed}

\item{isDuration}{Is duration of infusion measured (TRUE) or rate (FALSE)}

\item{isSequential}{Is model part of a PK/PD model that is being fitted sequentially}

\item{isClosedForm}{Is model closed-form algebraic (TRUE) or differential equation (FALSE)}
}
\description{
Class represents an NLME/PML PK model parameters
}
\examples{
      NlmePkParameters(ERR_ADDITIVE)

}
