% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{class}
\name{NlmePmlModel-class}
\alias{NlmePmlModel-class}
\alias{NlmePmlModel}
\title{NlmePmlModel}
\arguments{
\item{isLinearFrozen}{Is linear model frozen}

\item{userDefinedExtraDefs}{Custom definition for extra column
and table generation}
}
\description{
Class represents an NLME/PML model
}
\details{
individual (FALSE)
}
\examples{
      NlmePmlModel()

}
