% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure.r
\docType{class}
\name{NlmeProcedure-class}
\alias{NlmeProcedure-class}
\alias{NlmeProcedure}
\title{NlmeProcedure}
\arguments{
\item{structuralParams}{List of structural parameters to define}

\item{codeLines}{Lines of PML code to define continuous
function, including differential equations 
and logical statement}
}
\description{
Class represents a procedure block
}
\examples{
     Creates a procedure block and adds it to the object model

     proc = NlmeProcedure(structuralParams=list("Cl","C"),
                          codeLines=list("deriv(A1=Ka*Aa -Cl*C" ))
     addCompartment(model) = proc

}
