% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r, R/random_eff.r
\docType{class}
\name{NlmeRandomEffectBlock-class}
\alias{NlmeRandomEffectBlock-class}
\alias{NlmeRandomEffectBlock}
\title{XimportFrom utils View write.table
XimportFrom Certara.NLME8 NlmeDataset NlmeEngineExtraParams NlmeParallelHost}
\arguments{
\item{type}{Diagonal or Block}

\item{effectNames}{Names of random effects}

\item{frozen}{Are random effects frozen?}
}
\description{
Random effect block definition
}
\examples{
      NlmeRandomEffectBlock(Diagonal,list("Cl","V"),FALSE)

}
