% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_eff.r
\docType{class}
\name{NlmeRandomEffectValues-class}
\alias{NlmeRandomEffectValues-class}
\alias{NlmeRandomEffectValues}
\title{NlmeRandomEffectValues}
\arguments{
\item{numEffects}{Number of random effects}

\item{effectNames}{Names of random effects}

\item{values}{Matrix of random effect values}
}
\description{
Random effect values matrix
}
\examples{
      NlmeRandomEffectValues(2,c("Cl","V"),c(0.1,0,0,0,0,0.1))

}
