% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\docType{class}
\name{NlmeRandParamsMapping-class}
\alias{NlmeRandParamsMapping-class}
\alias{NlmeRandParamsMapping}
\title{NlmeRandParamsMapping}
\description{
Class represents mapping list between model variables and Random Effect
}
\examples{
   mapping = NlmeRandParamsMapping("list")

}
