% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\docType{class}
\name{NlmeResidualEffect-class}
\alias{NlmeResidualEffect-class}
\alias{NlmeResidualEffect}
\title{NlmeResidualEffect}
\arguments{
\item{effectName}{Name of the observed variable}

\item{observeName}{Name to use for observation}

\item{epsilonName}{Name to use for Epsilon}

\item{errorType}{Additive|LogAdditive|Multiplicative
|AdditiveMultiplicative|MixRatio|Power|Custom}

\item{frozen}{Is the standard deviation frozen?(default:FALSE)}

\item{SD}{Standard deviation value}

\item{definition}{Definition of a custom type or Power value}

\item{isBQL}{Are there BQL values?(default:FALSE)}

\item{bqlStatic}{Value LLOQ value}

\item{dobefore}{Code to execute before the observation}

\item{doafter}{Code to execute after the observation}
}
\description{
Class represents an NLME/PML residual error model
}
\examples{

 effect = NlmeResidualEffect(Additive,"C")

 effect = NlmeResidualEffect(AdditiveMultiplicative,"E",isBQL=TRUE)

}
