% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_param.r
\docType{class}
\name{NlmeStrucParamStyle-class}
\alias{NlmeStrucParamStyle-class}
\alias{NlmeStrucParamStyle}
\title{NlmeStrucParamStyle}
\arguments{
\item{style}{Parameter style: 1=LogNormal, 2=Normal, 
3=Combination, 4=Log, 
5=Logit, 6=Custom}
}
\description{
Class represents style of structural parameter
}
\examples{
       NlmeStrucParamStyle(style="1")

}
