% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\docType{class}
\name{ObservationCategorical-class}
\alias{ObservationCategorical-class}
\alias{ObservationCategorical}
\title{ObservationCategorical}
\arguments{
\item{observationName}{Observation name}

\item{linkFunction}{Logit|Probit|LogLog|Cloglog|InvCustom}

\item{isInhibitory}{Is the array of offset expressions in ascending order}

\item{slope}{Slope expression}

\item{numOutcomes}{Number of outcomes}

\item{offsetArray}{Offset expressions}

\item{dobefore}{Code to execute before effect is observed}

\item{doafter}{Code to execute after effect is observed}

\item{customCodeLines}{Lines of PML code to insert before LL() statement}
}
\description{
Class represents an NLME Categorical observation
}
\examples{

}
