% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\docType{class}
\name{ObservationContinuous-class}
\alias{ObservationContinuous-class}
\alias{ObservationContinuous}
\title{ObservationContinuous}
\arguments{
\item{name}{Observation name}

\item{residualEffect}{Defined by NlmeResidualEffect}

\item{customCodeLines}{Lines of PML code to insert before LL() statement}
}
\description{
Class represents an NLME continuous observation
}
\examples{

}
