% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\docType{class}
\name{ObservationLL-class}
\alias{ObservationLL-class}
\alias{ObservationLL}
\title{ObservationLL}
\arguments{
\item{observationName}{Observation name}

\item{slopeExpression}{Hazard slope expression}

\item{expression}{Base hazard expression}

\item{dobefore}{Code to execute before effect is observed}

\item{doafter}{Code to execute after effect is observed}

\item{simulationCode}{Lines of simulation code}

\item{customCodeLines}{Lines of PML code to insert before LL() statement}
}
\description{
Class represents an NLME LL observation
}
\examples{

}
