% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_param.r
\name{occasionCovariate}
\alias{occasionCovariate}
\title{occasionCovariate}
\usage{
occasionCovariate(name = "", occasions = c(), occasionNames = c(),
  direction = Backward, isDiagonal = TRUE)
}
\arguments{
\item{name}{Name of the covariate}

\item{occasions}{Numerical values for each occasion}

\item{occasionNames}{Names of each occasion}

\item{direction}{Curve fitting method: Forward|Backward|Interpolate}

\item{isDiagonal}{Is the structure diagonal (TRUE) or box (FALSE)}
}
\description{
Convenience method to create an occasion covariate
}
\examples{
      occ=occasionCovariate("OCC",
                        c(1,2),
                        c("OCC1","OCC2"),
                        direction=Forward)

}
