% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{parallelFirstOrderAbsorption}
\alias{parallelFirstOrderAbsorption}
\title{parallelFirstOrderAbsorption}
\usage{
parallelFirstOrderAbsorption(model, doseInfo1, doseInfo2)
}
\arguments{
\item{model}{PK/PD model to apply absorption model to}

\item{doseInfo}{Dose information for central compartment}

\item{absDoseInfo}{Dose information for absorption compartment}
}
\description{
Creates a parallel first order absorption dosing regimen
for the given model per dosing instructions
}
\examples{
      dosePoint1 = DosePoint(isZeroOrderAbsorption = NoDose,
                             isTlag = TRUE,
                             tlagExpression="Tlag",
                             isBioavail = TRUE,
                             bioavailExpression = "logitF1;1-ilogit(logitF1)")

      dosePoint2 = DosePoint(isZeroOrderAbsorption = NoDose,
                             isBioavail = TRUE,
                             bioavailExpression = "logitF1;ilogit(logitF1)")

      model2=parallelFirstOrderAbsorption(model,dosePoint1,dosePoint2)
      print(model2)

}
