% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_compartment.r
\docType{class}
\name{PDCompartment-class}
\alias{PDCompartment-class}
\alias{PDCompartment}
\title{PDCompartment}
\arguments{
\item{name}{Name of the PD compartment}
}
\description{
Class represents an NLME PD compartment
}
\examples{

}
