% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.r
\name{peripheralCompartment}
\alias{peripheralCompartment}
\title{peripheralCompartment}
\usage{
peripheralCompartment(..., aName = "", vName = "", cName = "",
  isConcentration = FALSE, isSequential = FALSE, numDosePoints = 0,
  compIndex = 1, blockName = "")
}
\arguments{
\item{aName}{Absorption variable name}

\item{vName}{Volume variable name}

\item{cName}{Observation variable name}

\item{isConcentration}{Are there volume and concentration parameters}

\item{isSequential}{Is compartment part of a sequential model}

\item{numDosePoints}{Number of dose points}

\item{compIndex}{Number to append to variable names identifier}

\item{blockName}{Name of the structural block}
}
\description{
Creates a peripheral compartment
}
\examples{
    periComp = peripheralCompartment(aName="Abs1",
                       vName="Vol",
                       cName="Conc",
                       isConcentration=TRUE,
                       isSequential=FALSE,
                       numDosePoints="2",
                       blockName="Peri1")

}
