% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{phxmodel}
\alias{phxmodel}
\title{Method to create a model from existing files generated by phoenix}
\usage{
phxmodel(modelName = "", dataset)
}
\arguments{
\item{modelName}{Visual name for the model}

\item{dataset}{Dataset representing phoenix generated model files}
}
\description{
Method to create a model from existing files generated by phoenix
}
\examples{

      dataset = NlmeDataset(phoenixSourceDir="C:/PhoenixFiles/PkModel")
      model = phxmodel(modelName="Test",
                       dataset = dataset )
      job = fitmodel(host,dataset)

}
