% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{pkemaxmodel}
\alias{pkemaxmodel}
\title{pkemaxmodel}
\usage{
pkemaxmodel(isPopulation = TRUE, parameterization = Clearance,
  absorption = Intravenous, numCompartments = 1, isClosedForm = TRUE,
  isTlag = FALSE, isTimeBased = TRUE, hasEliminationComp = FALSE,
  isFractionExcreted = FALSE, infusionAllowed = FALSE,
  isSaturating = FALSE, isDuration = FALSE, isSequential = FALSE,
  isPkFrozen = FALSE, hasEffectsCompartment = FALSE,
  checkBaseline = FALSE, checkFractional = FALSE,
  checkInhibitory = FALSE, checkSigmoid = FALSE,
  isEmaxFrozen = FALSE, covariates = c(), modelName = "",
  workingDir = "")
}
\arguments{
\item{isPopulation}{Population vs Individual models}

\item{parameterization}{Options are Clearance|Micro|Macro|Macro1}

\item{absorption}{Options are Extravascular|Intravenous}

\item{numCompartments}{1|2|3}

\item{isTlag}{Is there a dose time lag}

\item{isTimeBased}{Is model time-based}

\item{hasEliminationComp}{Is there an elimination compartment}

\item{isFractionExcreted}{Is elimination compartment fraction excreted}

\item{infusionAllowed}{Are infusions allowed}

\item{isSaturating}{Use Michaelis-Menten kinetics}

\item{isDuration}{Do infusions use duration instead of rate}

\item{isSequential}{Does PK model receive parameters from another model}

\item{isPkFrozen}{Is PK model frozen}

\item{hasEffectsCompartment}{Model has effect compartment}

\item{checkBaseline}{Does EMAX have a baseline}

\item{checkFractional}{Is EMAX fractional}

\item{checkInhibitory}{Is EMAX inhibitory}

\item{checkSigmoid}{Is EMAX sigmoid}

\item{isEmaxFrozen}{Is Emax model frozen}

\item{covariates}{Covariates defined for the model}

\item{modelName}{Visual name for the model}

\item{workingDir}{Working directory to run the model}
}
\description{
Method to create a PK/PD model
}
\examples{
      model=pkemaxmodel()

       OR

      model=pkemaxmodel( isPopulation=TRUE,
                           parameterization=Clearance,
                           absorption= ExtraVascular,
                           numCompartments=1,
                           isTlag=TRUE,
                           hasEliminationComp=TRUE,
                           hasEffectsComp=TRUE,
                           checkBaseline=TRUE))

      print(attr(model,"statements"))
      test(){
            cfMicro(A1,Cl/V, first = (Aa = Ka))
            dosepoint(Aa, tlag = Tlag)
            C = A1 / V
            E = E0 + Emax * C / (EC50 + C)
            error(CEps=1)
            observe(CObs=C + CEps)
            error(EEps=1)
            observe(EObs=E + EEps)
            stparm(Ka = tvKa * exp(nKa))
            stparm(V = tvV * exp(nV))
            stparm(Cl = tvCl * exp(nCl))
            stparm(Tlag = tvTlag * exp(nTlag))
            stparm(EC50 = tvEC50 * exp(nEC50))
            stparm(E0 = tvE0 * exp(nE0))
            stparm(Emax = tvEmax * exp(nEmax))
            fixef( tvKa = c(,1,))
            fixef( tvV = c(,1,))
            fixef( tvCl = c(,1,))
            fixef( tvTlag = c(,1,))
            fixef( tvEC50 = c(,1,))
            fixef( tvE0 = c(,1,))
            fixef( tvEmax = c(,1,))
            ranef(diag(nKa,nV,nCl,nTlag,nEC50,nE0,nEmax) =  c(1,1,1,1,1,1,1))
       }

}
