% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.r
\docType{class}
\name{PkFlow-class}
\alias{PkFlow-class}
\alias{PkFlow}
\title{PkFlow}
\arguments{
\item{c1Name}{Concentration variable name}

\item{flowName}{Name of flow}

\item{k12Name}{Name for the backward PK flow}

\item{k21Name}{Name for the forward flow}

\item{kmName}{Name for the fractional metabolic rate}

\item{vMaxName}{Name for the maximum  metabolic rate}

\item{flowType}{1=FpMicro, 2=FpClearance, 3=FpSaturating}

\item{fromPort}{Identifier of the source object's port}

\item{toPort}{Identifier of the target object's port}

\item{derivedVascularClearance}{Name for the clearance variable}

\item{twoWay}{Is the flow two-way (TRUE) or one-way (FALSE)}

\item{isVascular}{Is the flow vascular}

\item{fromReceptorPort}{}
}
\description{
Class represents PK flow
}
\examples{
      
}
