% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{pkindirectmodel}
\alias{pkindirectmodel}
\title{pkindirectmodel}
\usage{
pkindirectmodel(isPopulation = TRUE, parameterization = Clearance,
  absorption = Intravenous, numCompartments = 1, isClosedForm = TRUE,
  isTlag = FALSE, isTimeBased = TRUE, hasEliminationComp = FALSE,
  isFractionExcreted = FALSE, infusionAllowed = FALSE,
  isSaturating = FALSE, isDuration = FALSE, isSequential = FALSE,
  isPkFrozen = FALSE, indirectType = LIMITED_STIM,
  hasEffectsCompartment = FALSE, isBuildup = TRUE,
  isExponent = FALSE, indirectFrozen = FALSE, covariates = c(),
  modelName = "", workingDir = "")
}
\arguments{
\item{isPopulation}{Population vs Individual models}

\item{parameterization}{Options are Clearance|Micro|Macro|Macro1}

\item{absorption}{Options are Extravascular|Intravenous}

\item{numCompartments}{1|2|3}

\item{isTlag}{Is there a dose time lag}

\item{isTimeBased}{Is model time-based}

\item{hasEliminationComp}{Is there an elimination compartment}

\item{isFractionExcreted}{Is elimination compartment fraction excreted}

\item{infusionAllowed}{Are infusions allowed}

\item{isSaturating}{Use Michaelis-Menten kinetics}

\item{isDuration}{Do infusions use duration instead of rate}

\item{isSequential}{Does PK model receive parameters from another model}

\item{isPkFrozen}{Is PK model frozen}

\item{indirectType}{Type of model: LimitedStimulation|
InfiniteStimulation|LimitedInhibition|
InverseInhibition|LinearStimulation|
LogLinearStimulation}

\item{hasEffectsCompartment}{Model has effect compartment}

\item{isBuildup}{Do concentration changes cause a Buildup (TRUE)
or Loss (FALSE)}

\item{isExponent}{Is an exponent required}

\item{indirectFrozen}{Is Indirect model frozen}

\item{covariates}{List of covariates defined for the model}

\item{modelName}{Visual name for the model}

\item{workingDir}{Working directory to run the model}
}
\description{
Method to create a PK/Indirect PD model
}
\examples{
      pkpdmodel = pkindirectmodel(numComp=2,
                           absorption = Intravenous,
                           isClosedForm = FALSE,
                           isSequential = FALSE  ,
                           hasEliminationComp = FALSE,
                           indirectType=LimitedStimulation,
                           isBuildup=TRUE,
                           hasEffectsCompartment = TRUE,
                           modelName="PkIndirectModel")

      print(attr(model,"statements"))
      test(){
          deriv(A1 = - Cl * C - Cl2 * (C - C2))
          deriv(A2 = Cl2 * (C - C2))
          deriv(Ce = Ke0*(C - Ce))
          deriv(E = Kin * (1 + Emax * Ce / (Ce + EC50)) - Kout * E)
          dosepoint(A1)
          C = A1 / V
          C2 = A2 / V2
          sequence{E = Kin / Kout}
          error(CEps = 1)
          observe(CObs = C + CEps)
          error(EEps = 1)
          observe(EObs = E + EEps)
          stparm(V = tvV * exp(nV))
          stparm(V2 = tvV2 * exp(nV2))
          stparm(Cl = tvCl * exp(nCl))
          stparm(Cl2 = tvCl2 * exp(nCl2))
          stparm(Ke0 = tvKe0 * exp(nKe0))
          stparm(Kin = tvKin * exp(nKin))
          stparm(Kout = tvKout * exp(nKout))
          stparm(Emax = tvEmax * exp(nEmax))
          stparm(EC50 = tvEC50 * exp(nEC50))
          fcovariate(Age)
          fcovariate(BodyWeight)
          fcovariate(Gendercode())
          fixef(tvV = c(, 15.3918, ))
          fixef(tvV2 = c(, 41.2013, ))
          fixef(tvCl = c(, 6.60942, ))
          fixef(tvCl2 = c(, 14.038, ))
          fixef(tvKe0 = c(, 1, ))
          fixef(tvKin = c(, 1, ))
          fixef(tvKout = c(, 1, ))
          fixef(tvEmax = c(, 1, ))
          fixef(tvEC50 = c(, 1, ))
          ranef(diag(nV, nCl, nV2, nCl2, nKin, nKout,
                     nEmax, nEC50, nKe0) = c(0.06931719,
                     0.18120092, 3.0838168E-09,
                     0.042620177, 1, 1, 1, 1, 1))
          }

}
