% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{pkmodel}
\alias{pkmodel}
\title{pkmodel}
\usage{
pkmodel(isPopulation = TRUE, parameterization = Clearance,
  absorption = Intravenous, numCompartments = 1, isClosedForm = TRUE,
  isTlag = FALSE, isTimeBased = TRUE, hasEliminationComp = FALSE,
  isFractionExcreted = FALSE, infusionAllowed = FALSE,
  isSaturating = FALSE, isDuration = FALSE, isSequential = FALSE,
  covariates = c(), modelName = "", workingDir = "")
}
\arguments{
\item{isPopulation}{Population vs Individual models}

\item{parameterization}{Options are Clearance|Micro|Macro|Macro1}

\item{absorption}{Options are Extravascular|Intravenous}

\item{numCompartments}{1|2|3}

\item{isTlag}{Is there a dose time lag}

\item{isTimeBased}{Is model time-based}

\item{hasEliminationComp}{Is there an elimination compartment}

\item{isFractionExcreted}{Is elimination compartment fraction excreted}

\item{infusionAllowed}{Are infusions allowed}

\item{isSaturating}{Use Michaelis-Menten kinetics}

\item{isDuration}{Do infusions use duration instead of rate}

\item{isSequential}{Does PK model receive parameters from another model}

\item{covariates}{Covariates defined for the model}

\item{modelName}{Visual name for the model}

\item{workingDir}{Working directory to run the model}
}
\description{
Method to create a PK model
}
\examples{
      model = pkmodel(numCompartments=2)

      OR

      sex=categoricalCovariate("sex",c(0,1),c("male","female"))
      weight=NlmeCovariateParameter("weight",centerValue=75,
            continuousType=COVAR_MEDIAN)

      model=pkmodel( isPopulation=TRUE,
                           parameterization=Clearance,
                           absorption= Intravenous,
                           numCompartments=1,
                           isClosedForm=TRUE,
                           isTlag=FALSE,
                           isTimeBased=TRUE,
                           hasEliminationComp=FALSE,
                           isFractionExcreted=FALSE,
                           isSaturating=FALSE,
                           infusionAllowed=FALSE,
                           isDuration=FALSE,
                           isSequential=FALSE,
                           covariates=c(sex,weight))

      print(attr(model,"statements"))
      test(){
          cfMicro(A1, Cl / V, first = (Aa = Ka))
          dosepoint(Aa)
          C = A1 / V
          error(CEps = 0.0942729)
          observe(CObs = C * (1 + CEps))
          stparm(Ka = tvKa * exp(nKa))
          stparm(V = tvV * (wt/75)^dVdwt * exp(dVdsex1*(sex==1)) * exp(nV))
          stparm(Cl = tvCl * exp(nCl))
          covariate(sex())
          covariate(wt)
          fixef(tvKa = c(, 1, ))
          fixef(tvV = c(, 9.6764, ))
          fixef(tvCl = c(, 1, ))
          fixef(dVdwt(enable=c(0)) = c(, 0, ))
          fixef(dVdsex1(enable=c(1)) = c(, 0.101632, ))
          ranef(diag(nV, nCl, nKa) = c(0.10608772, 0.082715079, 1))
      }

}
