% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_compartment.r
\name{print.EffectsCompartment}
\alias{print.EffectsCompartment}
\title{Prints Effects compartment information}
\usage{
\method{print}{EffectsCompartment}(x, ...)
}
\arguments{
\item{obj}{Model compartment}
}
\description{
Prints Effects compartment information
}
\examples{
      print.EffectsCompartment(obj="linear")

}
