% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_compartment.r
\name{print.EmaxCompartment}
\alias{print.EmaxCompartment}
\title{print.EmaxCompartment}
\usage{
\method{print}{EmaxCompartment}(x, ...)
}
\arguments{
\item{obj}{Model compartment}
}
\description{
Prints Emax compartment information
}
\examples{
      print.EmaxCompartment(obj="linear")

}
