% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing.r
\name{print.ExtraDoseDataType}
\alias{print.ExtraDoseDataType}
\title{print.ExtraDoseDataType}
\usage{
\method{print}{ExtraDoseDataType}(x, ...)

\method{print}{ExtraDoseDataType}(x, ...)
}
\description{
print.ExtraDoseDataType

print.ExtraDoseDataType
}
