% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_indirect.r
\name{print.IndirectCompartment}
\alias{print.IndirectCompartment}
\title{print.IndirectCompartment}
\usage{
\method{print}{IndirectCompartment}(x, ...)
}
\arguments{
\item{obj}{Model compartment}
}
\description{
Prints indirect compartment information
}
\examples{
      print.IndirectCompartment(obj="linear")

}
