% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_linear.r
\name{print.LinearCompartment}
\alias{print.LinearCompartment}
\title{print.LinearCompartment}
\usage{
\method{print}{LinearCompartment}(x, ...)
}
\arguments{
\item{obj}{Model compartment}
}
\description{
Prints linear compartment information
}
\examples{
      print.LinearCompartment(obj="linear")

}
