% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{print.NlmeColumnMap}
\alias{print.NlmeColumnMap}
\title{print.NlmeColumnMap}
\usage{
\method{print}{NlmeColumnMap}(x, ...)
}
\arguments{
\item{obj}{Model object}
}
\description{
Prints column mapping
}
\examples{
   print.NlmeColumnMap(NlmeColumnMapping(mapping))

}
