% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.r
\name{print.NlmeCompartment}
\alias{print.NlmeCompartment}
\title{print.NlmeCompartment}
\usage{
\method{print}{NlmeCompartment}(x, ...)
}
\arguments{
\item{obj}{Model compartment}
}
\description{
Prints compartment information
}
\examples{
      print.NlmeCompartment(obj="Abs1")

}
