% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_param.r
\name{print.NlmeCovariateParameter}
\alias{print.NlmeCovariateParameter}
\title{print.NlmeCovariateParameter}
\usage{
\method{print}{NlmeCovariateParameter}(x, ...)
}
\arguments{
\item{obj}{Model covariate}
}
\description{
Prints covariate parameter information
}
