% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\name{print.NlmeErrorModel}
\alias{print.NlmeErrorModel}
\title{print.NlmeErrorModel}
\usage{
\method{print}{NlmeErrorModel}(x, ...)
}
\arguments{
\item{obj}{Error model name}
}
\description{
Prints Error model information
}
\examples{
      print.NlmeErrorModel(obj="C2")

}
