% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.r
\name{print.NlmeExpression}
\alias{print.NlmeExpression}
\title{print.NlmeExpression}
\usage{
\method{print}{NlmeExpression}(x, ...)
}
\arguments{
\item{obj}{Expression block name}
}
\description{
Prints Expression block information
}
\examples{
      print.NlmeExpression(obj="C2")

}
