% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure.r
\name{print.NlmeProcedure}
\alias{print.NlmeProcedure}
\title{print.NlmeProcedure}
\usage{
\method{print}{NlmeProcedure}(x, ...)
}
\arguments{
\item{obj}{Model compartment}
}
\description{
Prints procedure block information
}
\examples{
      print.NlmeProcedure(obj="proc")

}
