% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\name{print.NlmeResidualEffect}
\alias{print.NlmeResidualEffect}
\title{print.NlmeResidualEffect}
\usage{
\method{print}{NlmeResidualEffect}(x, ...)
}
\arguments{
\item{obj}{Residual effect name}
}
\description{
Prints Residual effect information
}
\examples{
      print.NlmeResidualEffect(obj="C2")

}
