% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_param.r
\name{print.NlmeStructuralParameter}
\alias{print.NlmeStructuralParameter}
\title{print.NlmeStructuralParameter}
\usage{
\method{print}{NlmeStructuralParameter}(x, ...)
}
\arguments{
\item{obj}{Structural parameter}
}
\description{
Prints structural parameter information
}
\examples{
      print.NlmeStructuralParameter(obj="EC50")

}
