% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\name{print.ObservationCategorical}
\alias{print.ObservationCategorical}
\title{print.ObservationCategorical}
\usage{
\method{print}{ObservationCategorical}(x, ...)
}
\arguments{
\item{obj}{Categorical observation name}
}
\description{
Prints Categorical observation information
}
\examples{
      print.ObservationCategorical(obj="pain")

}
