% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\name{print.ObservationContinuous}
\alias{print.ObservationContinuous}
\title{print.ObservationContinuous}
\usage{
\method{print}{ObservationContinuous}(x, ...)
}
\arguments{
\item{obj}{Continuous observation name}
}
\description{
Prints Continuous observation information
}
\examples{
      print.ObservationContinuous(obj="Ex_Amt")

}
