% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\name{print.ObservationEventCount}
\alias{print.ObservationEventCount}
\title{print.ObservationEventCount}
\usage{
\method{print}{ObservationEventCount}(x, ...)
}
\arguments{
\item{obj}{Event/Count observation name}
}
\description{
Prints Event/Count observation information
}
\examples{
      print.ObservationEventCount(obj="seizure")

}
