% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.r
\name{print.ObservationLL}
\alias{print.ObservationLL}
\title{print.ObservationLL}
\usage{
\method{print}{ObservationLL}(x, ...)
}
\arguments{
\item{obj}{LL observation name}
}
\description{
Prints LL observation information
}
\examples{
      print.ObservationLL(obj="LL_Obs")

}
