% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.r
\name{print.PkFlow}
\alias{print.PkFlow}
\title{print.PkFlow}
\usage{
\method{print}{PkFlow}(x, ...)
}
\arguments{
\item{obj}{Model flow}
}
\description{
Prints PK Flow information
}
\examples{
      print.PkFlow(obj="flowA")

}
