% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_variable.r
\name{print.SecondaryParameter}
\alias{print.SecondaryParameter}
\title{print.SecondaryParameter}
\usage{
\method{print}{SecondaryParameter}(x, ...)
}
\arguments{
\item{obj}{Secondary parameter}
}
\description{
Prints secondary parameter information
}
\examples{
      print.SecondaryParameter(obj)

}
