% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.r
\name{print.Wire}
\alias{print.Wire}
\title{print.Wire}
\usage{
\method{print}{Wire}(x, ...)
}
\arguments{
\item{obj}{Model wire}
}
\description{
Prints Wire information
}
\examples{
      print.Wire(obj="WireA1")

}
