% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_eff.r
\name{randomBlockStatement}
\alias{randomBlockStatement}
\title{randomBlockStatement}
\usage{
randomBlockStatement(.Object)
}
\arguments{
\item{model}{PK/PD model}
}
\description{
Returns random block statement
}
