% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing.r
\docType{class}
\name{ResetColumnInfo-class}
\alias{ResetColumnInfo-class}
\alias{ResetColumnInfo}
\title{ResetColumnInfo}
\arguments{
\item{low}{Lower value of reset range}

\item{hi}{Upper value of reset range}
}
\description{
ResetColumnInfo
}
\examples{

}
