% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{resetCovariateEffects}
\alias{resetCovariateEffects}
\title{resetCovariateEffects}
\usage{
resetCovariateEffects(.Object)
}
\arguments{
\item{model}{PK/PD model}
}
\description{
Returns a new model with  all covariate effects cleared.
}
\examples{
      newModel = resetCovariateEffects(pkmodel)

      covariateEffect(newModel,"wt","Cl")=COVAR_EFF_YES

}
