% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{residualEffect}
\alias{residualEffect}
\title{residualEffect}
\usage{
residualEffect(.Object, effectName)
}
\arguments{
\item{effectName}{Effect to lookup/set attributes for}

\item{model}{Model}

\item{List}{of attributes to set}
}
\description{
Method for getting/setting residual effect attributes
}
\examples{
    residualEffect(model,"C")

    residualEffect(model,"C") = c(errorType=LogAdditive)

    residualEffect(model,"E") = c(errorType=LogAdditive,
                                  isBQL=TRUE,
                                  frozen=TRUE)

    residualEffect(model,"C") = c(errorType=Multiplicative,
                                  SD="0.094345")

    residualEffect(model,"C") = c(errorType=Power,
                                  SD="0.094345",
                                  definition="3")

    See ?NlmeResidualEffect help for all attributes for this object


}
