% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{saveModel}
\alias{saveModel}
\title{saveModel}
\usage{
saveModel(model, dataset = NULL, engine = NULL, host = NULL)
}
\arguments{
\item{model}{PK/PD model}

\item{dataset}{Optional dataset to save with model}

\item{engine}{Optional engine parameters to save with model}

\item{host}{Optional host to save with model}
}
\description{
Saves the model object in model's working directory
}
\examples{
      saveModel(pkmodel)

      model=loadModel(directoryPath)

}
