% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_variable.r
\docType{class}
\name{SecondaryParameter-class}
\alias{SecondaryParameter-class}
\alias{SecondaryParameter}
\title{SecondaryParameter}
\arguments{
\item{name}{Name of the secondary parameter}

\item{definition}{Definition of secondary variable}

\item{unit}{Optional units}
}
\description{
Class represents an NLME secondary parameter
}
\examples{
     param = SecondaryParameter("Spc_Param","log(2)/tvKe")

     param = SecondaryParameter("Tmax",
                                "CalcTMax(tvA,tvAlpha,tvB,tvBeta,C,Gamma)")

}
