% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{setModel}
\alias{setModel}
\title{setModel}
\usage{
setModel(model)
}
\arguments{
\item{model}{PK/PD model}
}
\description{
Changes the current working directory to model's directory
}
\examples{
      setModel(pkmodel)

}
