% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing.r
\docType{data}
\name{SteadyStateDose}
\alias{SteadyStateDose}
\title{Steady State Dose}
\format{An object of class \code{numeric} of length 1.}
\usage{
SteadyStateDose
}
\description{
Steady State Dose
}
\keyword{datasets}
