% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.r
\docType{class}
\name{StructuralBlock-class}
\alias{StructuralBlock-class}
\alias{StructuralBlock}
\title{StructuralBlock}
\arguments{
\item{blockName}{Name of the structural block}

\item{compKey}{Identifier for the block}

\item{isSequential}{Is structural block in a sequential model}
}
\description{
Class represents structural blocks
}
\examples{
      StructuralBlock(blockName="Abs1",compKey="-1",isSequential=FALSE)

}
