% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{structuralParam<-}
\alias{structuralParam<-}
\title{Method to set structural parameter attributes}
\usage{
structuralParam(.Object, parameterNames) <- value
}
\arguments{
\item{model}{Model with the parameter}

\item{name}{Name of the parameter}

\item{fixedEffName}{Name of fixed effect}

\item{randomEffName}{Name of random effect}

\item{style}{One of LogNormal|Normal|Combination|Log|Logit|Custom}

\item{initialValue}{Initial value of the fixed effect}

\item{lowerBound}{Lower limit value of fixed effect}

\item{upperBound}{Upper limit value of fixed effect}

\item{units}{Unit of measurement of fixed effect data}

\item{isFrozen}{Freeze fixed effect?( default FALSE )}

\item{hasRandomEffect}{Has random effect? (default TRUE)}

\item{code}{Line of code that defines stparm()}

\item{extraCode}{Extra lines of code that relates to this parameter}
}
\description{
Method to set structural parameter attributes
}
\examples{
     structuralParam(model,"Cl") = c(style=LogNormal,initialValue="0.75")

     structuralParam(model,"Cl2") = c(style=Custom, code="stparm(V=10^(tvlog10V + nlog10V))")

}
