% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{structuralParam}
\alias{structuralParam}
\title{Method to list attributes for a structural parameter}
\usage{
structuralParam(.Object, parameterName)
}
\arguments{
\item{model}{Model with the parameter}

\item{name}{Parameter name}
}
\description{
Method to list attributes for a structural parameter
}
\examples{
      structuralParam(model,"Cl")

}
