% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{updateDatasetDosing}
\alias{updateDatasetDosing}
\title{Formats column names of a data table and writes it out
to given file name}
\usage{
updateDatasetDosing(model, dataset)
}
\description{
Formats column names of a data table and writes it out
to given file name
}
