% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{updateDatasetParams}
\alias{updateDatasetParams}
\title{Formats column names of a data table and writes it out
to given file name}
\usage{
updateDatasetParams(model, dataset)
}
\description{
Formats column names of a data table and writes it out
to given file name
}
