% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{updateDatasetRandParams}
\alias{updateDatasetRandParams}
\title{Formats column names of a data table and writes it out
to given file name}
\usage{
updateDatasetRandParams(model, dataset)
}
\description{
Formats column names of a data table and writes it out
to given file name
}
