% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\docType{methods}
\name{updateErrorModel,NlmePmlModel-method}
\alias{updateErrorModel,NlmePmlModel-method}
\title{updateErrorModel}
\usage{
\S4method{updateErrorModel}{NlmePmlModel}(.Object)
}
\description{
Updates error model, recreating some structural parameters
that depend on how residual effects are used
}
