% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{userDefinedExtraDefinitions}
\alias{userDefinedExtraDefinitions}
\title{userDefinedExtraDefinitions}
\usage{
userDefinedExtraDefinitions(.Object, covariateName)
}
\arguments{
\item{model}{A PK/PD model}

\item{userDefinedList}{Lines of extra column/table definition}
}
\description{
Lists/sets user defined extra column/table definition
}
\examples{

    userDefinedExtraDefinitions(model)

    userDefinedExtraDefinitions(model) = c("addlcol(ADDL)",
           " iicol(II)", "table(file=\\"res.csv\\",time(0),Ka,V,Cl,Tlg)")

}
