% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.r
\docType{class}
\name{Wire-class}
\alias{Wire-class}
\alias{Wire}
\title{Wire}
\arguments{
\item{wireName}{Name of the wire connection}

\item{fromObject}{Identifier of the source object}

\item{fromPort}{Identifier of the source object's port}

\item{toObject}{Identifier of the target object}

\item{toPort}{Identifier of the target object's port}

\item{contains}{}
}
\description{
Class represents flow connection between compartments
}
\examples{
      
}
