% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{workspace}
\alias{workspace}
\title{workspace}
\usage{
workspace(directory = "")
}
\arguments{
\item{directory}{Workspace root directory}
}
\description{
Sets the given directory as root of where models get created OR
switches cwd to the previously set workspace
}
\examples{
      workspace(directory)

      workspace()

}
