% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{writeColumnMapping}
\alias{writeColumnMapping}
\title{writeColumnMapping}
\usage{
writeColumnMapping(model, filename)
}
\arguments{
\item{model}{A PK/PD model}

\item{filename}{Name of the file}
}
\description{
Write the mapping between model variables and column names to a file
}
\examples{
   writeColumnMapping(model,filename)

}
