% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{writeDefaultFiles}
\alias{writeDefaultFiles}
\title{writeDefaultFiles}
\usage{
writeDefaultFiles(model, dataset, simParams = NULL)
}
\arguments{
\item{model}{PK/PD model}

\item{dataset}{Dataset to run simulation/fitting on}

\item{simParams}{Simulation Parameters}
}
\description{
Writes out data/column mapping and engine parameter files for the engine
}
