% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{writeDoseMapping}
\alias{writeDoseMapping}
\title{writeDoseMapping}
\usage{
writeDoseMapping(model, dataset)
}
\arguments{
\item{model}{A PK/PD model}

\item{dataset}{Source of dose data}
}
\description{
Write the mapping between model variables and dose column names to a file
}
\examples{
   writeDoseMapping(model,dataset)

}
