% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{writeInputData}
\alias{writeInputData}
\title{writeInputData}
\usage{
writeInputData(model, datafileName)
}
\arguments{
\item{model}{A PK/PD model}

\item{datafileName}{Input data file}
}
\description{
Write the input data for the model to a file
}
\examples{
   writeDoseMapping(model,datafileName)

}
