% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{writeParamsMapping}
\alias{writeParamsMapping}
\title{writeParamsMapping}
\usage{
writeParamsMapping(model, dataset)
}
\arguments{
\item{model}{A PK/PD model}

\item{dataset}{Source of parameter data}
}
\description{
Write the mapping between model variables and parameter column
names to a file
}
\examples{
   writeParamsMapping(model,dataset)

}
