% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.r
\name{writeRandParamsMapping}
\alias{writeRandParamsMapping}
\title{writeRandParamsMapping}
\usage{
writeRandParamsMapping(model, dataset)
}
\arguments{
\item{model}{A PK/PD model}

\item{dataset}{Source of random effect data}
}
\description{
Write the mapping between model variables and random effect column
names to a file
}
\examples{
   writeRandParamsMapping(model,dataset)

}
