% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_model.r
\name{zeroOrderAbsorption}
\alias{zeroOrderAbsorption}
\title{zeroOrderAbsorption}
\usage{
zeroOrderAbsorption(model, doseInfo)
}
\arguments{
\item{model}{PK/PD model to apply absorption model to}

\item{doseInfo}{Dose information for central compartment}
}
\description{
Creates a zeroth order absorption dosing regimen for the model
}
\examples{
      dosePoint= DosePoint(isZeroOrderAbsorption = DurationDose,
                           durationExpression = "DUR",
                           isTlag = TRUE,
                           tlagExpression = "Tlag")

      model2=zeroOrderAbsorption(model,dosePoint)
      print(model2)

}
